Refactored the forms to keep project organized and more readable
