+++ client/src/Components/login/index.js
@@ -1,6 +1,7 @@
-import React, { Component }from 'react'
-import { Formik, Form, Field, ErrorMessage } from 'formik'
-import { LogInValidation } from '../../Validator/ValidationSchema'
+import React, { Component } from 'react'
+import { Formik } from 'formik'
+import { LogInValidation } from '../../Forms/ValidationSchema'
+import { LogInForm } from '../../Forms/Forms'
 import axios from 'axios'
 import './styles.css'
 
@@ -31,33 +32,8 @@ class LogIn extends Component {
                 }}
                 validationSchema={ LogInValidation }
                 onSubmit={ this.handleSubmit }
-                render={ ({ errors, touched, handleSubmit }) => (
-                    <Form>
-                        <div className="fieldDiv">
-                            <Field
-                                className="field"
-                                type="email"
-                                name="email"
-                                style={{ borderColor: errors.email && touched.email ? 'red' : null }}
-                                placeholder="Enter your school email"
-                            />
-                            <ErrorMessage component="p" name="email" />
-                        </div>
-                        <div className="fieldDiv">
-                            <Field
-                                className="field"
-                                type="password"
-                                name="password"
-                                style={{ borderColor: errors.password && touched.password ? 'red' : null }}
-                                placeholder="Enter password"
-                            />
-                            <ErrorMessage component="p" name="password" />
-                        </div>
-                        <br /><br />
-                        <div className="buttonDiv">
-                            <button id="submitBtn" type="submit" onClick={handleSubmit}>Submit</button>
-                        </div>
-                    </Form>
+                render={ props => (
+                    <LogInForm {...props}/>
                 )}
             />
         )
